Fix: Compilation error with GCC-14.2 on Ubuntu 224.04
This change fixes sign-conversion warning that is escalated to an error
when building using GCC-14.2 from Ubuntu 24.04. This is exact platform
that is used for API documentation builder.

There might be a cleaner approach and more sign conversion could be
addressed in this file, but this is outside of the scope of this change
as it is only aimed to resolve an immediate failures on the CI.

+++ source/blender/blenlib/BLI_linear_allocator.hh
@@ -244,7 +244,7 @@ template<typename Allocator = GuardedAllocator> class LinearAllocator : NonCopya
     /* If the original allocation size was large, it might have been separately allocated. In this
      * case, we can't free the end of it anymore. */
     if (original_allocation_size <= large_buffer_threshold) {
-      const int64_t new_begin = uintptr_t(free_after);
+      const int64_t new_begin = int64_t(uintptr_t(free_after));
       BLI_assert(new_begin <= current_begin_);
 #ifndef NDEBUG
       /* This condition is not really necessary but it helps finding the cases where memory was
